Extract Claims in a separate directory,create an index file
