+++ ts/claims/index.ts
@@ -1,0 +1,7 @@
+export default class Claims {
+  
+  
+  public createVerifiableCredential(issuer, credentialType, claim, privateKeyWIF) {
+
+  }
+}
