+++ ts/identity/types.ts
@@ -29,12 +29,3 @@ export interface IpfsConfig {
   port: number;
   protocol: string;
 }
-
-export interface IVerifiableCredentialAttrs {
-  '@context': string;
-  id: string;
-  "type": string[];
-  issuer: string;
-  issued: string;
-  claim: { id: string; [x:string]:any };
-}
